+++ source/Programs/Compositor/GodRays/raysMod.frag
@@ -16,7 +16,7 @@
 uniform sampler2D RT;
 uniform sampler2D FBO;
 
-in mediump vec2 vUV0;
+in highp vec2 vUV0;
 void main()
 {
     mediump vec3 color = texture2D(RT, vUV0).rgb;
