+++ source/Programs/Compositor/HDR/tonemap.frag
@@ -15,7 +15,7 @@
 uniform sampler2D RT;
 
 // https://github.com/OGRECave/ogre-next/blob/v2.3.1/Samples/Media/2.0/scripts/materials/HDR/GLSL/FinalToneMapping_ps.glsl
-in mediump vec2 vUV0;
+in highp vec2 vUV0;
 void main()
 {
     mediump vec3 color = texture2D(RT, vUV0).rgb;
