+++ source/Programs/Compositor/SMAA/SMAA_EdgeDetection.frag
@@ -29,7 +29,7 @@ uniform mediump vec4 ViewportSize;
 
 #include "smaa.glsl"
 
-in mediump vec2 vUV0;
+in highp vec2 vUV0;
 in mediump mat4 offset;
 void main()
 {
